+++ src/kimi_cli/tools/file/write.py
@@ -9,8 +9,8 @@ from kimi_cli.soul.agent import BuiltinSystemPromptArgs
 from kimi_cli.soul.approval import Approval
 from kimi_cli.tools.display import DisplayBlock
 from kimi_cli.tools.file import FileActions
-from kimi_cli.tools.file.utils import build_diff_blocks
 from kimi_cli.tools.utils import ToolRejectedError, load_desc
+from kimi_cli.utils.diff import build_diff_blocks
 from kimi_cli.utils.path import is_within_directory